Bugfix: mechanism to keep the keybind "command string" was broken
This is an really old bug :) Guess no one used the functionality.

* Background

ioncore.getbindings() returns all bindings in a structured format.

The action is/was however only available as a lua function, makeing it
very hard to decode what the binding actually does.

Digging in the source we find that this actually is unintended and a
mechanism is in place to keep the string representation of the actions.

A simple bug prevented this from working.

Seems like notion stores the "command string" in a weak hashmap in
"lua land" instead of in the core data structure in "c land".

ioncore.getbindings then augments "c land" datastructure with the
command strings on demand.

Fixing this bug make it possible to create better keybind help.