Don't enforced "floating" mode on frames managed by WGroup.
This fixes a bug where loaded layouts doesn't preserve the frame mode. (Frames
managed by WGroupWS with tiled-alt mode (bar mode) where reverted to floating on
restart)

Not really sure when the removed code is needed. I guess if one where to move a
whole frame from a tiled workspace to an empty workspace the code would change
the mode. Not sure when that is actually done though.

It might be possible to detect whether we're in a "load-from-layout" situation
or not, but didn't see any obvious signs.